refactor(database): separate database.ts by schema
